requirements.txt

streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0
requests>=2.31.0